MDL-79593 ltixservice_gradebookservices: migrate tables during install
This is the second and final step in migrating the data from the now-
defunct ltiservice_gradebookservices plugin. This data was moved to a
temporary table during core upgrade so as to sidestep plugin uninstall
code. This data is now migrated into the replacement plugin and the tmp
table dropped.