MDL-75947 ltiservice_gradebookservices: test covering lineitem rescale
This test highlights a problem in which updates to the scoreMaximum will
trigger a grade rescale, resulting in the timemodified value for grades
being set to the current time. Any subsequent score posts to that line
item must set a timestamp greater than this time to be considered valid.

This test will currently fail because the logic is too strict. A tool
should be able to post a subsequent score update, so long as the time
stamp is greater than the most recent update. Since the tool is unaware
of the grade rescale event having taken place, or the requirements to
increase the score timestamp to a value higher than the grade rescale
time, this then causes problems during post, resulting in 409s.

Solutions:
1. It's not as simple as removing the date check logic, as this will
likely result in certification suite failure.
2. We might need to track the last score post time in the grade service
plugin, so that we can check against that, instead of relying on the
grade->timemodified value, which clearly can change for other reasons.
3. Data fix won't be possible for existing grades.

We need to look into this problem further.