MDL-84415 core_ltix: expand resource_link definition
Adds the usual trifecta of fields used to link to a component:
{component, itemtype, itemid}, as well as a boolean 'gradable' field.

This gradable field is useful because, in future,  not all links will be
in an activity context like mod_lti, and thus won't all have associated
lineitem claims which need to be included during launch. This flag will
help service plugins make quick decisions about what claims to include
based on whether the link supports grades, such as a link owned by
mod_lti, or doesn't (such as a tool link placed inside an editor, or
in course navigation).