MDL-87128 mod_lti: refactor mod/lti/view to use launch container helper
The mod/lti/view.php file now fetches the resource link for that $lti
instance, and uses that instead of the $lti instance when asking for
the launch container. This commit also improves code organisation,
and adds some project TODOs, noting down a few pieces of code that can
potentially be removed in future.