MDL-84653 core_ltix: Update privacy provider to use resource links
The core_ltix privacy provider is no longer coupled to an LTI external
tool ID when it acts as a subsystem that provides data to other
components (for example, to obtain or remove LTI submission data).
It now uses resource links to help components retrieve, update, or
delete the relevant user data.
In addition to this update, minor refactoring was carried out as part
of the change.